# Conflicts:
#	CHANGELOG.md
#	cnccode/controller_classes/CTIgnoredADDomains.php
#	cnccode/controller_classes/CTOffice365Licenses.inc.php
#	cnccode/controller_classes/CTUtilityEmail.inc.php
#	dbVersioning/db.changelog-master.xml
#	htdocs/components/package.json
#	htdocs/components/services/APIMain.js
#	htdocs/components/services/ApiUrls.js

+++ htdocs/components/services/APIMain.js
@@ -1,7 +1,7 @@
 class APIMain {
     get(url) {
         return fetch(url)
-            .then((res) => res.json());
+        .then((res) => this.handleResponse(res));
     }
 
     getCustomers() {
@@ -58,10 +58,15 @@ class APIMain {
             }).then((res) => this.handleResponse(res));
     }
 
-    delete(url) {
+    delete(url,handleException=false) {
+      if (!handleException)
         return fetch(url, {
             method: "DELETE",
         }).then((res) => res.json());
+        else
+        return fetch(url, {
+          method: "DELETE",
+      }).then((res) => this.handleResponse(res));
     }
 
     postFormData(url, payload) {
